requirements.txt
Flask
SQLAlchemy
Flask-SQLAlchemy
gunicorn
Pillow
